requirements.txt
fastapi == 0.85.0
loguru == 0.6.0
uvicorn == 0.18.3
jinja2 == 3.1.2
black == 22.8.0
flake8 == 5.0.4
SQLAlchemy == 1.4.41
psycopg2 == 2.9.3
python-decouple == 3.6
requests == 2.28.1